API/requirements.txt
azure.storage.blob>=12.12.0
torch<=1.10.1
pydantic<=1.9.1
fastapi<=0.78.0
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
jsonpickle<=2.2.0
numpy<=1.22.4
Pillow<=9.1.1
opencv-python-headless<=4.5.5.64
pandas<=1.4.2
seaborn<=0.11.2
PyYAML<=6.0
matplotlib<=3.5.2
tqdm<=4.64.0
torchaudio<=0.10.1
torchvision<=0.11.2
ipython<=8.3.0
psutil<=5.9.1